# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MultiTrack Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}/pay:
    post:
      operationId: payOrder
      tags: [orders]
      summary: Initiate payment for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayOrderBody"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"

  # ── Downloads ─────────────────────────────────────────────────────────────
...
  /downloads/{grantId}/link:
    post:
      operationId: generateDownloadLink
      tags: [downloads]
      summary: Generate a temporary signed download link
      parameters:
        - name: grantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadLink"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Subscriptions ─────────────────────────────────────────────────────────
...
  /rateios:
    get:
      operationId: listRateios
      tags: [rateios]
      summary: List rateios
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [suggested, in_quotation, open, goal_reached, awaiting_payment, payment_confirmed, in_progress, completed, cancelled, refunded]
        - name: genre
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rateio"
    post:
      operationId: suggestRateio
      tags: [rateios]
      summary: Suggest a new rateio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestRateioBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rateio"
  /rateios/{id}:
    get:
      operationId: getRateio
      tags: [rateios]
      summary: Get rateio details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateioDetail"
  /rateios/{id}/join:
    post:
      operationId: joinRateio
      tags: [rateios]
      summary: Join a rateio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateioParticipant"
  /rateios/{id}/comments:
    get:
      operationId: getRateioComments
      tags: [rateios]
      summary: Get rateio comments/updates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateioComment"
    post:
      operationId: addRateioComment
      tags: [rateios]
      summary: Add comment to rateio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateioComment"

  # ── Wishlist ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isBlocked:
          type: boolean
        ordersCount:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, isBlocked, ordersCount, totalSpent, createdAt]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        artist:
          type: string
        genre:
          type: string
        coverUrl:
          type: string
          nullable: true
        price:
          type: number
        promoPrice:
          type: number
          nullable: true
        bpm:
          type: integer
          nullable: true
        tonality:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        quality:
          type: string
          enum: [premium, standard, backing_track, demo]
        status:
          type: string
          enum: [active, inactive, draft, featured]
        isFeatured:
          type: boolean
        creditsRequired:
          type: integer
        availableForSubscription:
          type: boolean
        totalSales:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, artist, genre, price, quality, status, isFeatured, creditsRequired, availableForSubscription, totalSales, createdAt]
...
    CreateOrderBody:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [pix, credit_card]
        couponCode:
          type: string
          nullable: true
      required: [paymentMethod]
    PayOrderBody:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [pix, credit_card]
        cardToken:
          type: string
          nullable: true
      required: [paymentMethod]
    PaymentResponse:
      type: object
      properties:
        orderId:
          type: integer
        status:
          type: string
        pixCode:
          type: string
          nullable: true
        pixQrCode:
          type: string
          nullable: true
        checkoutUrl:
          type: string
          nullable: true
      required: [orderId, status]
    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, paid, cancelled, refunded, chargeback]
        paymentMethod:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        couponCode:
          type: string
          nullable: true
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, status, paymentMethod, subtotal, discount, total, itemCount, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        price:
          type: number
        downloadGrant:
          $ref: "#/components/schemas/DownloadGrant"
          nullable: true
      required: [id, product, price]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            user:
              $ref: "#/components/schemas/UserSummary"

    # ── Downloads ─────────────────────────────────────────────────────────
    DownloadGrant:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productArtist:
          type: string
        coverUrl:
          type: string
          nullable: true
        source:
          type: string
          enum: [purchase, credit, rateio]
        downloadCount:
          type: integer
        lastDownloadAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, productId, productName, productArtist, source, downloadCount, createdAt]
    DownloadLink:
      type: object
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [url, expiresAt]

    # ── Subscriptions ─────────────────────────────────────────────────────
...
    SuggestRateioBody:
      type: object
      properties:
        songName:
          type: string
        artist:
          type: string
        reference:
          type: string
          nullable: true
        justification:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
      required: [songName, artist]
    Rateio:
      type: object
      properties:
        id:
          type: integer
        songName:
          type: string
        artist:
          type: string
        genre:
          type: string
          nullable: true
        status:
          type: string
          enum: [suggested, in_quotation, open, goal_reached, awaiting_payment, payment_confirmed, in_progress, completed, cancelled, refunded]
        targetAmount:
          type: number
          nullable: true
        amountPerParticipant:
          type: number
          nullable: true
        minParticipants:
          type: integer
          nullable: true
        currentParticipants:
          type: integer
        minPercentage:
          type: number
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        coverUrl:
          type: string
          nullable: true
        isParticipating:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, songName, artist, status, currentParticipants, isParticipating, createdAt]
    RateioParticipant:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        paymentStatus:
          type: string
          enum: [pending, paid, overdue, refunded]
        joinedAt:
          type: string
          format: date-time
      required: [id, userId, userName, paymentStatus, joinedAt]
    RateioComment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        isAdmin:
          type: boolean
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, userName, isAdmin, content, createdAt]
    RateioDetail:
      allOf:
        - $ref: "#/components/schemas/Rateio"
        - type: object
          properties:
            justification:
              type: string
              nullable: true
            participants:
              type: array
              items:
                $ref: "#/components/schemas/RateioParticipant"
            comments:
              type: array
              items:
                $ref: "#/components/schemas/RateioComment"
            product:
              $ref: "#/components/schemas/Product"
              nullable: true
    AddCommentBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...